FixStage

+++ Assets/Member/HJH/02.Scripts/Customer/CustomerType.cs
@@ -3,15 +3,15 @@ using UnityEngine;
 
 [CreateAssetMenu(fileName = "CustomerType", menuName = "H_SO/CustomerType")]
 
-public class CustomerType : ScriptableObject
-{
-    public CustomerCategory customerCategory;
-    public int customerDamage;
-    public float attackInterval;
-    public int customerHP;
-    public float customerSpeed;
-    public int money;
-}
+    public class CustomerType : ScriptableObject
+    {
+        public CustomerCategory customerCategory;
+        public int customerDamage;
+        public float attackInterval;
+        public float customerHP;
+        public float customerSpeed;
+        public int money;
+    }
 public enum CustomerCategory
 {
     person,
